requirements.txt
# 核心依赖
datasets>=2.14.0
transformers>=4.35.0
huggingface-hub>=0.19.0

# 数据处理
pyyaml>=6.0
jsonlines>=4.0.0
tqdm>=4.65.0

# 分析和可视化
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Lean4 解析（可选）
# lark-parser>=1.1.0  # 用于语法解析

# 工具
colorlog>=6.7.0
